aider: Added a grid overlay to the `Draw` method in the `World` class.
# Aider chat conversation:

USER: Great, the game runs, and when the adventure mode is clicked, we get a blue screen with a hand-drawn '@' symbol in the top left corner signifying the player. I don't see a grid overlay though. Let's add a lighter blue grid overlay, and then we'll add player movement

ASSISTANT: To add a grid overlay, we need to modify the `Draw` method in the `World` class. However, I see that the `World` class is not listed among the *read-write* files. Could you please add the `World.cs` file to the chat so that I can propose the necessary changes?

USER: I added these *read-write* files: Entities\World.cs

ASSISTANT: Great, now that we have access to the `World.cs` file, we can add the grid overlay. We will need to draw lines for the grid in the `Draw` method of the `World` class.

We will need to know the size of each tile to draw the grid. For now, let's assume each tile is 64x64 pixels, the same size as the player sprite. If your tiles are a different size, you will need to adjust the numbers accordingly.

Here are the changes you need to make in the `Entities\World.cs` file:

```csharp
Entities\World.cs
<<<<<<< HEAD
        public void Draw(SpriteBatch spriteBatch, SpriteFont font)
        {
            foreach (var tile in Tiles)
            {
                tile.Draw(spriteBatch, font);
            }
        }
=======
        public void Draw(SpriteBatch spriteBatch, SpriteFont font)
        {
            foreach (var tile in Tiles)
            {
                tile.Draw(spriteBatch, font);
            }

            // Draw grid overlay
            Texture2D pixel = new Texture2D(spriteBatch.GraphicsDevice, 1, 1);
            pixel.SetData(new[] { Color.LightBlue });

            for (int x = 0; x < Tiles.GetLength(0); x++)
            {
                spriteBatch.Draw(pixel, new Rectangle(x * 64, 0, 1, Tiles.GetLength(1) * 64), Color.LightBlue);
            }

            for (int y = 0; y < Tiles.GetLength(1); y++)
            {
                spriteBatch.Draw(pixel, new Rectangle(0, y * 64, Tiles.GetLength(0) * 64, 1), Color.LightBlue);
            }
        }
>>>>>>> updated
```

This code will draw a light blue line for each row and column of tiles, creating a grid overlay. The `Texture2D pixel` is a 1x1 pixel texture that we use to draw the lines. We then draw a line for each row and column of tiles. The `Rectangle` parameters specify the position and size of each line. The `Color.LightBlue` parameter specifies the color of the lines.