aider: Added SpriteBatchExtensions.cs with a DrawLine method for drawing lines using a SpriteBatch.
# Aider chat conversation:

USER: We are getting Hex and Layout namespace errors in world.cs now, as well as SpriteBatch not containing DrawLine

ASSISTANT: It seems like the `Hex` and `Layout` classes are not recognized, and the `DrawLine` method is not found in the `SpriteBatch` class. This could be due to missing references or the classes/methods not being implemented.

Here's what we need to do:

1. Ensure that the `Hex` and `Layout` classes are implemented and are in the correct namespace. If they are in a different namespace, you might need to add a `using` statement at the top of the `World.cs` file.

2. The `DrawLine` method is not a built-in method of the `SpriteBatch` class. It's a common extension method provided by some libraries. If you don't have such a method, you'll need to implement it.

Here's a simple implementation of the `DrawLine` method as an extension method for the `SpriteBatch` class. This should be added to a new file in the `Utils` directory:

```csharp
Utils\SpriteBatchExtensions.cs
<<<<<<< HEAD
=======
using Microsoft.Xna.Framework;
using Microsoft.Xna.Framework.Graphics;

namespace EpochsUnbound.Utils
{
    public static class SpriteBatchExtensions
    {
        public static void DrawLine(this SpriteBatch spriteBatch, Point start, Point end, Texture2D texture, Color color)
        {
            Vector2 edge = end.ToVector2() - start.ToVector2();
            float angle = (float)Math.Atan2(edge.Y, edge.X);

            spriteBatch.Draw(texture,
                new Rectangle((int)start.x, (int)start.y, (int)edge.Length(), 1),
                null,
                color,
                angle,
                new Vector2(0, 0),
                SpriteEffects.None,
                0);
        }
    }
}
>>>>>>> updated
```

This extension method allows you to draw a line from a start point to an end point using a `SpriteBatch`. It calculates the angle of the line, then draws a 1-pixel wide rectangle from the start point to the end point at the calculated angle.

After adding this extension method, the `DrawLine` method should be available to use in the `World.cs` file.