aider: Changed player and world drawing to use the EpochsUnbound.Utils.Point class instead of Microsoft.Xna.Framework.Point.
# Aider chat conversation:

USER: now 'Point' is an ambiguous reference

ASSISTANT: The error message "Point is an ambiguous reference" typically means that there are two or more classes named `Point` in the namespaces that you're using, and the compiler doesn't know which one you mean.

In your case, it seems like there's a conflict between `EpochsUnbound.Utils.Point` and `Microsoft.Xna.Framework.Point`.

To resolve this issue, you can use fully qualified names for `Point` in your code. This means specifying the namespace along with the class name every time you use `Point`.

Here are the proposed changes:

```csharp
Entities\World.cs
<<<<<<< HEAD
            spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new Microsoft.Xna.Framework.Point(64, 64)), Color.White);
=======
            spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new EpochsUnbound.Utils.Point(64, 64)), Color.White);
>>>>>>> updated
```

```csharp
Entities\World.cs
<<<<<<< HEAD
            Layout layout = new Layout(Layout.FlatTop, new Point(64, 64), new Point(0, 0));
=======
            Layout layout = new Layout(Layout.FlatTop, new EpochsUnbound.Utils.Point(64, 64), new EpochsUnbound.Utils.Point(0, 0));
>>>>>>> updated
```

```csharp
Entities\World.cs
<<<<<<< HEAD
            spriteBatch.DrawLine(start, end, pixel, Color.LightBlue);
=======
            spriteBatch.DrawLine(start.ToVector2(), end.ToVector2(), pixel, Color.LightBlue);
>>>>>>> updated
```

Please apply these changes and let me know if the issue is resolved. If not, we may need to look into the implementation of the `Point` class and its usage in other parts of your code.

USER: InvalidEditBlock: edit failed!

Entities\World.cs does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in Entities\World.cs with nothing missing!

Entities\World.cs does not contain these 1 exact lines in a row:
```
            spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new Microsoft.Xna.Framework.Point(64, 64)), Color.White);
```

ASSISTANT: I apologize for the confusion. It seems I made a mistake in identifying the correct file and line of code to modify. The line of code I referenced is not in the `Entities\World.cs` file.

Let's correct this. The changes should be applied to the `Entities\Player.cs` file instead. Here is the correct edit block:

```csharp
Entities\Player.cs
<<<<<<< HEAD
            spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new Microsoft.Xna.Framework.Point(64, 64)), Color.White);
=======
            spriteBatch.Draw(playerTexture, new Rectangle(Position.ToPoint(), new EpochsUnbound.Utils.Point(64, 64)), Color.White);
>>>>>>> updated
```

And for the `Entities\World.cs` file, the changes should be:

```csharp
Entities\World.cs
<<<<<<< HEAD
            Layout layout = new Layout(Layout.FlatTop, new Point(64, 64), new Point(0, 0));
=======
            Layout layout = new Layout(Layout.FlatTop, new EpochsUnbound.Utils.Point(64, 64), new EpochsUnbound.Utils.Point(0, 0));
>>>>>>> updated
```

```csharp
Entities\World.cs
<<<<<<< HEAD
            spriteBatch.DrawLine(start, end, pixel, Color.LightBlue);
=======
            spriteBatch.DrawLine(start.ToVector2(), end.ToVector2(), pixel, Color.LightBlue);
>>>>>>> updated
```

Please apply these changes and let me know if the issue is resolved. If not, we may need to look into the implementation of the `Point` class and its usage in other parts of your code.