aider: Changed the third argument of the `Layout` constructor to be of type `EpochsUnbound.Utils.Point` instead of `Microsoft.Xna.Framework.Point`.
# Aider chat conversation:

USER: Argument 3: cannot convert from 'Microsoft.Xna.Framework.Point' to 'EpochsUnbound.Utils.Point'

ASSISTANT: The error message indicates that there is a type mismatch between `Microsoft.Xna.Framework.Point` and `EpochsUnbound.Utils.Point`. This is likely happening because the `Layout` constructor is expecting a `EpochsUnbound.Utils.Point` but is being provided with a `Microsoft.Xna.Framework.Point`.

To fix this, we need to ensure that the third argument passed to the `Layout` constructor is of type `EpochsUnbound.Utils.Point`.

Here's the change:

```csharp
Entities\World.cs
<<<<<<< HEAD
    Layout layout = new Layout(Layout.FlatTop, new EpochsUnbound.Utils.Point(64, 64), new Microsoft.Xna.Framework.Point(0, 0));
=======
    Layout layout = new Layout(Layout.FlatTop, new EpochsUnbound.Utils.Point(64, 64), new EpochsUnbound.Utils.Point(0, 0));
>>>>>>> updated
```

This change will ensure that the third argument passed to the `Layout` constructor is of type `EpochsUnbound.Utils.Point`, resolving the issue.