aider: Added 'ToVector2' method to 'Point' class in 'Utils/Point.cs'.
# Aider chat conversation:

USER: 'Point' does not cointain a definition for 'ToVector2'

ASSISTANT: The error message indicates that the `Point` class from the `EpochsUnbound.Utils` namespace does not have a method named `ToVector2`.

To fix this, we need to add a `ToVector2` method to the `Point` class. This method will convert the `Point` to a `Vector2`.

Here's the change:

```csharp
Utils\Point.cs
<<<<<<< HEAD
    public struct Point
    {
        public double x;
        public double y;

        public Point(double x, double y)
        {
            this.x = x;
            this.y = y;
        }
    }
=======
    public struct Point
    {
        public double x;
        public double y;

        public Point(double x, double y)
        {
            this.x = x;
            this.y = y;
        }

        public Vector2 ToVector2()
        {
            return new Vector2((float)x, (float)y);
        }
    }
>>>>>>> updated
```

This change will add a `ToVector2` method to the `Point` class, resolving the issue.