aider: Added using directive for Microsoft.Xna.Framework namespace in Point.cs.
# Aider chat conversation:

USER: The type or namespace name 'Vector2' could not be found

ASSISTANT: The error message indicates that the `Vector2` class is not recognized. This is likely because the namespace that contains the `Vector2` class is not being used in the file where the error occurs.

The `Vector2` class is part of the `Microsoft.Xna.Framework` namespace. To fix this issue, we need to add a `using` directive for this namespace at the top of the file where the error occurs.

However, without knowing the exact file where the error occurs, I can't provide a specific solution. Could you please specify the file where this error is occurring?

USER: the error is in point.cs

ASSISTANT: The `Vector2` class is part of the `Microsoft.Xna.Framework` namespace. To fix this issue, we need to add a `using` directive for this namespace at the top of the `Point.cs` file.

Here's the change:

```csharp
Utils\Point.cs
<<<<<<< HEAD
namespace EpochsUnbound.Utils
{
    public struct Point
    {
        public double x;
        public double y;

        public Point(double x, double y)
        {
            this.x = x;
            this.y = y;
        }

        public Vector2 ToVector2()
        {
            return new Vector2((float)x, (float)y);
        }
    }
}
=======
using Microsoft.Xna.Framework;

namespace EpochsUnbound.Utils
{
    public struct Point
    {
        public double x;
        public double y;

        public Point(double x, double y)
        {
            this.x = x;
            this.y = y;
        }

        public Vector2 ToVector2()
        {
            return new Vector2((float)x, (float)y);
        }
    }
}
>>>>>>> updated
```

This change will add a `using` directive for the `Microsoft.Xna.Framework` namespace at the top of the `Point.cs` file, resolving the issue.