aider: Refactored `DrawLine` method to accept arguments of type `Vector2` instead of `Epochs.utils.point`.
# Aider chat conversation:

USER: Argument 2: cannot convert from 'Ms.xna.framework.vector2' to 'Epochs.utils.point'

ASSISTANT: The error message "Argument 2: cannot convert from 'Ms.xna.framework.vector2' to 'Epochs.utils.point'" suggests that the `DrawLine` method in the `SpriteBatchExtensions` class is expecting arguments of type `Epochs.utils.point`, but it's receiving arguments of type `Ms.xna.framework.vector2`.

To resolve this issue, we need to modify the `DrawLine` method in the `SpriteBatchExtensions` class to accept arguments of type `Ms.xna.framework.vector2` and convert them to `Epochs.utils.point` inside the method.

Here is the proposed change:

```csharp
Utils\SpriteBatchExtensions.cs
<<<<<<< HEAD
    public static void DrawLine(this SpriteBatch spriteBatch, EpochsUnbound.Utils.Point start, EpochsUnbound.Utils.Point end, Texture2D texture, Color color)
    {
        Vector2 startVector = new Vector2((float)start.x, (float)start.y);
        Vector2 endVector = new Vector2((float)end.x, (float)end.y);
        Vector2 edge = endVector - startVector;
        float angle = (float)Math.Atan2(edge.Y, edge.X);

        spriteBatch.Draw(texture,
            new Rectangle((int)startVector.X, (int)startVector.Y, (int)edge.Length(), 1),
            null,
            color,
            angle,
            new Vector2(0, 0),
            SpriteEffects.None,
            0);
    }
=======
    public static void DrawLine(this SpriteBatch spriteBatch, Vector2 start, Vector2 end, Texture2D texture, Color color)
    {
        Vector2 edge = end - start;
        float angle = (float)Math.Atan2(edge.Y, edge.X);

        spriteBatch.Draw(texture,
            new Rectangle((int)start.X, (int)start.Y, (int)edge.Length(), 1),
            null,
            color,
            angle,
            new Vector2(0, 0),
            SpriteEffects.None,
            0);
    }
>>>>>>> updated
```

This code modifies the `DrawLine` method to accept arguments of type `Vector2` instead of `Epochs.utils.point`. It also updates the calculation of the `startVector` and `endVector` variables to use the `X` and `Y` properties of the `Vector2` class instead of the `x` and `y` properties of the `Epochs.utils.point` class.